Update: #231 Wave 2 후속 — android.graphics.Color.parseColor → commonMain parseHexColor 통일
