+++ feature/memo-plain/impl/src/commonMain/kotlin/me/pecos/memozy/presentation/screen/memo/TextStyleModel.kt
@@ -105,27 +105,3 @@ fun toggleStyle(
 
     return result.sortedBy { it.start }
 }
-
-private fun parseHexColor(hex: String): Color {
-    return try {
-        val cleaned = hex.removePrefix("#")
-        when (cleaned.length) {
-            6 -> {
-                val r = cleaned.substring(0, 2).toInt(16)
-                val g = cleaned.substring(2, 4).toInt(16)
-                val b = cleaned.substring(4, 6).toInt(16)
-                Color(red = r, green = g, blue = b)
-            }
-            8 -> {
-                val a = cleaned.substring(0, 2).toInt(16)
-                val r = cleaned.substring(2, 4).toInt(16)
-                val g = cleaned.substring(4, 6).toInt(16)
-                val b = cleaned.substring(6, 8).toInt(16)
-                Color(red = r, green = g, blue = b, alpha = a)
-            }
-            else -> Color.Unspecified
-        }
-    } catch (_: Exception) {
-        Color.Unspecified
-    }
-}
